+++ lib/src/internal/table_row_widget.dart
@@ -70,6 +70,7 @@ class TableRowWidget<ROW> extends StatelessWidget {
         painter: TableRowPainter<ROW>(
             row: row,
             cellPadding: theme.cell.padding,
+            cellAlignment: theme.cell.alignment,
             columns: columns,
             columnsMetrics: columnsMetrics,
             contentHeight: contentHeight,